Simplified custom example
